refactor: modularize hooks and improve validation persistence logic
