+++ app/components/CodeEditor/CodeEditor.tsx
@@ -7,130 +7,20 @@ import Editor, { Monaco } from "@monaco-editor/react";
 import { Flex, useColorMode } from "@chakra-ui/react";
 import { useEffect, useState, useRef, useCallback } from "react";
 import MyBtn from "../MyBtn";
-import {
-  tryFormattingCode,
-  validateCode,
-  restorePreviousValidation,
-  hasValidationResult,
-} from "@/lib/client-functions";
+import { tryFormattingCode, validateCode } from "@/lib/client-functions";
 import FiChevronRight from "@/app/styles/icons/HiChevronRightGreen";
 import { useRouter } from "next/navigation";
-import { useUserSolutionStore, useEditorStore } from "@/lib/stores";
+import { useEditorStore } from "@/lib/stores";
 import { sendGAEvent } from "@next/third-parties/google";
 import { CodeFile, OutputResult } from "@/lib/types";
 import { OutputReducerAction } from "@/lib/reducers";
 import CertificateButton from "../CertificateButton/CertificateButton";
-
-// Custom hook for editor theme setup
-const useEditorTheme = (monaco: Monaco, colorMode: "dark" | "light") => {
-  useEffect(() => {
-    if (monaco) {
-      monaco.editor.defineTheme("my-theme", {
-        base: "vs-dark",
-        inherit: true,
-        rules: [],
-        colors: {
-          "editor.background": "#1f1f1f",
-        },
-      });
-      monaco.editor.setTheme(colorMode === "light" ? "light" : "my-theme");
-    }
-  }, [monaco, colorMode]);
-};
-
-// Custom hook for keyboard shortcuts
-const useValidationShortcut = (
-  handleValidate: () => void,
-  codeString: string,
-) => {
-  useEffect(() => {
-    const handleKeyDown = (event: KeyboardEvent) => {
-      if (event.key === "Enter" && event.shiftKey) {
-        sendGAEvent("event", "buttonClicked", {
-          value: "Validate (through shortcut)",
-        });
-        event.preventDefault();
-        handleValidate();
-      }
-    };
-
-    document.addEventListener("keydown", handleKeyDown);
-    return () => {
-      document.removeEventListener("keydown", handleKeyDown);
-    };
-  }, [handleValidate, codeString]);
-};
-
-// Custom hook for code persistence
-const useCodePersistence = (
-  chapterIndex: number,
-  stepIndex: number,
-  codeString: string,
-  setCodeString: (value: string) => void,
-  codeFile: CodeFile,
-) => {
-  const userSolutionStore = useUserSolutionStore();
-
-  // Load saved code
-  useEffect(() => {
-    const savedCode = userSolutionStore.getSavedUserSolutionByLesson(
-      chapterIndex,
-      stepIndex,
-    );
-    if (savedCode && savedCode !== codeString) {
-      setCodeString(savedCode);
-    }
-  }, [chapterIndex, stepIndex]);
-
-  // Save code changes
-  useEffect(() => {
-    userSolutionStore.saveUserSolutionForLesson(
-      chapterIndex,
-      stepIndex,
-      codeString,
-    );
-  }, [codeString, chapterIndex, stepIndex]);
-
-  // Initialize code if no saved solutions
-  useEffect(() => {
-    if (Object.keys(userSolutionStore.userSolutionsByLesson).length === 0) {
-      setCodeString(JSON.stringify(codeFile.code, null, 2));
-    }
-  }, [userSolutionStore]);
-};
-
-// Custom hook for validation restoration
-const useValidationRestore = (
-  chapterIndex: number,
-  stepIndex: number,
-  dispatchOutput: React.Dispatch<OutputReducerAction>,
-  setCodeString: (value: string) => void,
-) => {
-  const [isRestored, setIsRestored] = useState(false);
-
-  useEffect(() => {
-    // Restore previous validation on component mount or when lesson changes
-    if (!isRestored && hasValidationResult(chapterIndex, stepIndex)) {
-      try {
-        const { restored } = restorePreviousValidation(
-          chapterIndex, 
-          stepIndex, 
-          dispatchOutput, 
-          setCodeString
-        );
-        if (restored) {
-          setIsRestored(true);
-          console.log('✅ Previous validation restored for lesson:', chapterIndex, stepIndex);
-        }
-      } catch (error) {
-        console.error('Failed to restore validation:', error);
-      }
-    }
-  }, [chapterIndex, stepIndex, isRestored, dispatchOutput, setCodeString]);
-
-  return { isRestored };
-};
-
+import {
+  useEditorTheme,
+  useValidationShortcut,
+  useCodePersistence,
+  useValidationRestore,
+} from "@/app/utils/hooks";
 
 // EditorControls component for the buttons section
 const EditorControls = ({
@@ -230,7 +120,7 @@ export default function CodeEditor({
       );
       setIsValidating(false);
     }, 500);
-  }, [codeString, codeFile, dispatchOutput, stepIndex, chapterIndex]);
+  }, [codeString, codeFile, dispatchOutput, stepIndex, chapterIndex, setCodeString]);
 
   useValidationShortcut(handleValidate, codeString);
   useCodePersistence(
@@ -241,6 +131,7 @@ export default function CodeEditor({
     codeFile,
   );
 
+  // Restore previous validation on lesson revisit
   const { isRestored } = useValidationRestore(
     chapterIndex,
     stepIndex,
@@ -248,12 +139,13 @@ export default function CodeEditor({
     setCodeString,
   );
 
+  // Reset code to initial state
   const resetCode = () => {
     setCodeString(JSON.stringify(codeFile.code, null, 2));
     dispatchOutput({ type: "RESET" });
   };
 
-  const handleEditorMount = (editor: monaco.editor.IStandaloneCodeEditor, monacoInstance: Monaco) => {
+  const handleEditorMount = (editor: any, monacoInstance: Monaco) => {
     setMonaco(monacoInstance);
 
     editorRef.current = editor;
@@ -263,6 +155,7 @@ export default function CodeEditor({
 
   return (
     <>
+      {/* Show success banner when previous validation is restored */}
       {isRestored && (
         <div
           style={{
